requirements.txt
chromadb>=0.5.5,<1
fastapi>=0.115,<1
langchain>=0.3,<0.4
langchain-chroma>=0.2,<1
langchain-community>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-huggingface>=0.1,<1
langchain-ollama>=0.2,<1
langchain-openai>=0.3,<1
langchain-text-splitters>=0.3,<0.4
pydantic-settings>=2.5,<3
pypdf>=5,<6
python-dotenv>=1,<2
python-multipart>=0.0.12,<1
sentence-transformers>=3.0,<4
uvicorn[standard]>=0.31,<1
